splits dma into more modules
